Add Algolia search integration
This is admittedly overkill but I really wanted to try it, and well, why
not.

Judging by the redux docs and CRA, it's fine to put the API key in plain
text. Besides, it only allows other sides to search my own, and it's not
like I can obscure it behind a server anyway.